# docs/openapi/provisioner-api.yaml
openapi: 3.0.3
info:
  title: eapaka-node-provisioner API
  version: 0.3.0
  description: |
    EAP-AKA RADIUS PoC（以下「本PoC」）の Provisioning API（以下「prov」）と、aka-only-server の管理API（以下「aka」）を
    組み合わせて操作する統合API。設計は `docs/design-overview.md` を参照。
    作法は下流の 2 つの API（prov 0.4.0、aka 0.2.0）に揃えている。

    0.3.0 の変更:
    - 加入者の停止・再開（`PUT /policies/{imsi}/status`。prov の中継）を加えた。停止の印は本PoCの認可ポリシーの状態
      （`status`: `active` / `suspended`）で、鍵の置き場所によらず本PoCの Auth Server が認証を拒否する。
    - 加入者の統合リソースに `status`（認可ポリシーの状態）を加えた。作成・変更の本文では指定できない（作成は常に `active`）。

    加入者:
    - 加入者を「IMSI＋鍵の置き場所（`keyStore`）＋認可ポリシー」として扱う（`/subscribers`）。
      - `poc`（接続方式 `00`）: 鍵を本PoCの `sub:{IMSI}` に持つ（prov の `/subscribers`）。
      - `aka`（接続方式 `01`）: 鍵を aka-only-server が持つ（aka の `/subscribers`）。本PoCの vector-gateway の
        AVクライアントID（設定 `PROVISIONER_AKA_AV_CLIENT_ID`）を、加入者の許可クライアントに含める。
      - 認可ポリシーは、置き場所によらず本PoCに持つ（prov の `/policies`）。
    - 置き場所は、IMSI の PLMN と設定 `PROVISIONER_PLMN_MAP`（本PoCの `VECTOR_GATEWAY_PLMN_MAP` と同じ形式・同じ値）で決まる。
      先頭 6 桁、次に先頭 5 桁の順で照合し、`01` に当たれば `aka`、それ以外は `poc`。
    - 複数のノードにまたがる作成・変更・削除は、操作の記録（`/operations`）を残しながら行い、途中で失敗したら
      補償（作成・変更）またはやり直し（削除）で整える。

    中継:
    - RADIUSクライアント（`/clients`）、認可ポリシー（`/policies`）、セッション（`/sessions`）は prov と同じ形で中継する。
      要求の本文は検証せずにそのまま送り、下流の応答（エラーを含む）をそのまま返す。
    - 下流の監査ログ（`/prov/audit-logs`、`/aka/audit-logs`）と、vector-gateway の AVクライアント（`/aka/av-clients/{clientId}`）は、
      下流の形のまま返す（読み取りだけ）。

    認証:
...
    - 管理クライアントは全権として扱う。利用者や権限の概念はこの API にはない。
...
servers:
...
paths:
...
  /subscribers:
    parameters:
      - $ref: '#/components/parameters/OperatorId'
      - $ref: '#/components/parameters/TraceId'
    get:
      tags: [subscribers]
      summary: 加入者の一覧を取得する
      operationId: listSubscribers
      description: |
        prov の加入者、aka の加入者、prov の認可ポリシーの 3 つの一覧を、同じ `cursor` と `prefix` で取り寄せ、
        IMSI の昇順にマージして返す。Ki と OPc は含まない。
        - aka の加入者は、PLMN マップで `aka` に当たる IMSI だけを対象にする（aka-only-server には本PoCと関係のない加入者もいるため）。
        - 鍵だけ、ポリシーだけの IMSI も含める（`issues` で示す）。
        - 一覧では、`poc` の IMSI について aka 側の加入者は調べない（`KEY_IN_OTHER_STORE` は個別の取得でだけ分かる場合がある）。
        - 和集合の正確な件数は求めないため、`total` は返さない。
      parameters:
        - $ref: '#/components/parameters/ImsiPrefix'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: 加入者の一覧
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriberList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '502':
          $ref: '#/components/responses/DownstreamError'
        '503':
          $ref: '#/components/responses/DownstreamUnavailable'
    post:
      tags: [subscribers]
      summary: 加入者を作成する
      operationId: createSubscriber
      description: |
        次の順に行う。
        1. 入力を検証する（下流と同じ規則。ポリシーの規則を含む）。
        2. prov の加入者、aka の加入者（置き場所が `aka` なら有無、`poc` なら vector-gateway の AVクライアントID を許可しているもの）、
           prov の認可ポリシーの有無を確かめる。
           どれかがあれば 409（`SUBSCRIBER_ALREADY_EXISTS`。`conflicts` にある場所）。
        3. 操作の記録を作り、置き場所に加入者（鍵）を作成する。
        4. 認可ポリシーを PUT する。

        4 が失敗したら 3 で作った加入者を削除して戻し（補償）、最初の失敗のエラーに `operationId` と `rolledBack: true` を付けて返す。
        補償も失敗したら 500（`OPERATION_INCOMPLETE`）を返し、後で自動でやり直す（`/operations`）。
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscriberCreate'
            examples:
              poc:
                summary: 鍵を本PoCに置く加入者
                value:
                  imsi: '001010000000001'
                  ki: 465b5ce8b199b49faa5f0a2ee238a6bc
                  opc: cd63cb71954a9f4e48a5994e37a02baf
                  policy:
                    default: deny
                    rules:
                      - nasId: '*'
                        allowedSsids: [CORP-WIFI]
              aka:
                summary: 鍵を aka-only-server に置く加入者（PLMN マップで 00102 を 01 にした場合）
                value:
                  imsi: '001020000000001'
                  keyStore: aka
                  ki: 465b5ce8b199b49faa5f0a2ee238a6bc
                  opc: cd63cb71954a9f4e48a5994e37a02baf
                  sqnType: inc32
                  policy:
                    default: allow
                    rules: []
      responses:
        '201':
          description: 作成した加入者。Ki と OPc は含まない。
          headers:
            Location:
              description: 作成した加入者の URI（例 `/admin/v1/subscribers/001010000000001`）
              schema:
                type: string
            Idempotent-Replayed:
              $ref: '#/components/headers/IdempotentReplayed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscriber'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/SubscriberConflict'
        '422':
          $ref: '#/components/responses/IdempotencyKeyMismatch'
        '500':
          $ref: '#/components/responses/InternalErrorOrIncomplete'
        '502':
          $ref: '#/components/responses/DownstreamError'
        '503':
          $ref: '#/components/responses/DownstreamUnavailable'
  /subscribers/{imsi}:
    parameters:
      - $ref: '#/components/parameters/Imsi'
      - $ref: '#/components/parameters/OperatorId'
      - $ref: '#/components/parameters/TraceId'
    get:
      tags: [subscribers]
      summary: 加入者を取得する
      operationId: getSubscriber
      description: |
        prov の加入者、aka の加入者、prov の認可ポリシーを並行して取得し、まとめて返す。Ki と OPc は含まない。
        3 か所のどこにもなければ 404。食い違いは `issues` で示す。
        置き場所でない方の下流にだけ接続できない場合は、`issues` に `OTHER_STORE_UNREACHABLE` を入れて 200 を返す。
      responses:
        '200':
          description: 加入者
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscriber'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/SubscriberNotFound'
        '500':
          $ref: '#/components/responses/InternalError'
        '502':
          $ref: '#/components/responses/DownstreamError'
        '503':
          $ref: '#/components/responses/DownstreamUnavailable'
    patch:
      tags: [subscribers]
      summary: 加入者を変更する
      operationId: updateSubscriber
      description: |
        指定した項目だけを変更する（JSON Merge Patch）。次の順に行う。
        1. 変更前の認可ポリシーを操作の記録に残す（`policy` を指定したとき）。
        2. 認可ポリシーを PUT する（`policy` を指定したとき。全体を置き換える）。
        3. 置き場所の加入者を PATCH する（鍵の項目を指定したとき）。

        3 が失敗したら、2 を変更前に戻し（変更前になければ削除）、最初の失敗のエラーに `operationId` と `rolledBack: true` を付けて返す。
        戻せなければ 500（`OPERATION_INCOMPLETE`）を返し、後で自動でやり直す。

        - `sqn` を指定しなければ SQN には触れない（認証で進んだ SQN を巻き戻さない）。指定した場合は、そのまま書き換える。
        - `sqnType` / `allowPlain` は `keyStore` が `aka` の加入者だけに指定できる。許可クライアント（`allowedClientIds`）は変更しない。
        - 置き場所は変更できない。
        - 鍵の項目を指定したが置き場所に加入者がない場合は 404（`USER_NOT_FOUND`）。`policy` だけなら、鍵がなくても変更できる。
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: '#/components/schemas/SubscriberUpdate'
            example:
              amf: b9b9
              policy:
                default: deny
                rules:
                  - nasId: AP-OFFICE-01
                    allowedSsids: [CORP-WIFI]
                    vlanId: '100'
      responses:
        '200':
          description: 変更後の加入者。Ki と OPc は含まない。
          headers:
            Idempotent-Replayed:
              $ref: '#/components/headers/IdempotentReplayed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscriber'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/SubscriberNotFound'
        '409':
          $ref: '#/components/responses/OperationInProgress'
        '422':
          $ref: '#/components/responses/IdempotencyKeyMismatch'
        '500':
          $ref: '#/components/responses/InternalErrorOrIncomplete'
        '502':
          $ref: '#/components/responses/DownstreamError'
        '503':
          $ref: '#/components/responses/DownstreamUnavailable'
    delete:
      tags: [subscribers]
      summary: 加入者を削除する
      operationId: deleteSubscriber
      description: |
        認可ポリシー、置き場所の加入者（鍵）の順に削除する（どちらも、既にないものは削除済みとして扱う）。
        削除は戻せないため、途中で失敗したら 500（`OPERATION_INCOMPLETE`）を返し、残りを後で自動でやり直す。
        - 置き場所でない方の下流にある同じ IMSI の加入者（`KEY_IN_OTHER_STORE`）は削除しない。
        - `aka` の加入者は、他の AVクライアントを許可していても、加入者ごと削除する（他の AVクライアントとは共有しない前提）。
        - 同じ IMSI のセッションは削除しない（本PoCの Provisioning API と同じ）。
        - 3 か所のどこにもなければ 404。
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '204':
          description: 削除した
          headers:
            Idempotent-Replayed:
              $ref: '#/components/headers/IdempotentReplayed'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/SubscriberNotFound'
        '409':
          $ref: '#/components/responses/OperationInProgress'
        '422':
          $ref: '#/components/responses/IdempotencyKeyMismatch'
        '500':
          $ref: '#/components/responses/InternalErrorOrIncomplete'
        '502':
          $ref: '#/components/responses/DownstreamError'
        '503':
          $ref: '#/components/responses/DownstreamUnavailable'
  /subscribers/{imsi}/keys:
    parameters:
      - $ref: '#/components/parameters/Imsi'
      - $ref: '#/components/parameters/OperatorId'
      - $ref: '#/components/parameters/TraceId'
    get:
      tags: [subscribers]
      summary: 加入者の Ki と OPc を取得する
      operationId: getSubscriberKeys
      description: |
        置き場所の `/subscribers/{imsi}/keys` から取得する。provisioner と置き場所の下流の監査ログに記録する（値は記録しない）。
      responses:
        '200':
          description: Ki と OPc
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriberKeys'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/SubscriberNotFound'
        '500':
          $ref: '#/components/responses/InternalError'
        '502':
          $ref: '#/components/responses/DownstreamError'
        '503':
          $ref: '#/components/responses/DownstreamUnavailable'
...
components:
  parameters:
    Imsi:
      name: imsi
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Imsi'
...
    ImsiPrefix:
      name: prefix
      in: query
      description: IMSI の前方一致条件。
      schema:
        type: string
        pattern: '^[0-9]{1,15}$'
    Cursor:
      name: cursor
      in: query
      description: 前のページの `nextCursor`（IMSI）。省略時は先頭から返す。
      schema:
        type: string
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 500
        default: 50
...
    OperatorId:
      name: X-Operator-Id
      in: header
      required: false
      description: 操作者 ID。監査ログに記録し、下流にもそのまま渡す。形式が違えば 400（`OPTIONAL_IE_INCORRECT`）。
      schema:
        type: string
        pattern: '^[A-Za-z0-9._@-]{1,64}$'
    TraceId:
      name: X-Trace-ID
      in: header
      required: false
      description: |
        トレースID。ログ・監査ログ・操作の記録に残し、下流 2 つにも同じ値を渡す。省略時、または形式が違う（印字可能ASCII 1〜64文字でない）場合は
        provisioner が採番する（16進32桁）。使った値を応答の `X-Trace-ID` ヘッダーで返す（エラーの応答を含む）。
      schema:
        type: string
        pattern: '^[\x21-\x7E]{1,64}$'
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        再送の判定に使うキー（印字可能ASCII 1〜128文字）。同じ管理クライアントから同じキーで届いた要求には、
        24 時間のあいだ最初の応答を返す。形式が違えば 400（`OPTIONAL_IE_INCORRECT`）。
        5xx と、一時的な 409（`OPERATION_IN_PROGRESS`、`OPERATION_UNRESOLVED`）の応答は覚えない（同じキーで送り直せる）。
      schema:
        type: string
        pattern: '^[\x21-\x7E]{1,128}$'
  headers:
    IdempotentReplayed:
      description: '`Idempotency-Key` により、以前の応答をそのまま返した場合だけ `true`。'
      schema:
        type: string
        enum:
          - 'true'
  responses:
    BadRequest:
      description: |
        リクエストが不正。`cause` は次のいずれか。項目ごとの理由は `invalidParams` に入る。
        - `INVALID_MSG_FORMAT`: JSON として解釈できない、Content-Type が違う、未知の項目がある
        - `INVALID_QUERY_PARAM`: クエリパラメーターの値が不正
        - `MANDATORY_IE_MISSING`: 必須項目がない
        - `MANDATORY_IE_INCORRECT`: 必須項目の値が不正（パスの IMSI / ID を含む）
        - `OPTIONAL_IE_INCORRECT`: 任意項目の値が不正（PATCH の `null`、`X-Operator-Id` / `Idempotency-Key` の形式、
          `keyStore` が `poc` の加入者への `sqnType` / `allowPlain` を含む）
        - `KEY_STORE_MISMATCH`: 作成で指定した `keyStore` が、PLMN マップから決まる置き場所と違う（`invalidParams` の `param` は `keyStore`）

        下流が入力の誤りとして断った場合（provisioner の検証をすり抜けた場合）も、下流の `cause` のまま返す。
        `invalidParams` の `param` は provisioner の項目名に付け替える（例 `rules[0].vlanId` → `policy.rules[0].vlanId`）。
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    SubscriberNotFound:
      description: |
        加入者が存在しない。`cause` は `USER_NOT_FOUND`。
        - 取得・削除: 3 か所（prov の加入者、aka の加入者、ポリシー）のどこにもない。
        - 変更: 3 か所のどこにもない、または鍵の項目を指定したが置き場所に加入者がない。
        - 鍵の取得: 置き場所に加入者がない。
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    SubscriberConflict:
      description: |
        次のどちらか。
        - `SUBSCRIBER_ALREADY_EXISTS`: 同じ IMSI のものが既にある。`conflicts` にある場所を入れる
          （書き込みの途中で下流が 409 を返した場合は、補償したうえで `operationId` と `rolledBack` も付ける）。
        - `OPERATION_IN_PROGRESS`: 同じ IMSI の操作、または同じ `Idempotency-Key` の要求が処理中。
        - `OPERATION_UNRESOLVED`: 同じ IMSI に未完了の操作の記録がある。`operationId` にその操作の ID を入れる。
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
    IdempotencyKeyMismatch:
      description: 同じ `Idempotency-Key` で、内容（メソッド・パス・本文）の違う要求が届いた。`cause` は `IDEMPOTENCY_KEY_MISMATCH`。
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
    InternalError:
      description: provisioner 内部のエラー（Valkey のエラー等）。`cause` は `SYSTEM_FAILURE`。
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    InternalErrorOrIncomplete:
      description: |
        次のどちらか。
        - `OPERATION_INCOMPLETE`: 下流への書き込みの途中で失敗し、補償（作成・変更）または残りの削除も失敗した。
          `operationId` を付ける。provisioner が後で自動でやり直す（`/operations/{operationId}`）。
          `downstream` / `downstreamStatus` / `downstreamCause` は最初の失敗のもの。
        - `SYSTEM_FAILURE`: provisioner 内部のエラー（Valkey のエラー等）。Valkey に接続できない場合、下流には何もしない。
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    DownstreamError:
      description: |
        下流が想定外の応答（5xx、ProblemDetails でない応答、呼び出し側では直せない 4xx）を返した。`cause` は `DOWNSTREAM_ERROR`。
        `downstream`、`downstreamStatus`、分かれば `downstreamCause` を付ける。
        例: aka が `CLIENT_NOT_FOUND`（設定の `PROVISIONER_AKA_AV_CLIENT_ID` の AVクライアントがない）を返した。
        書き込みの途中なら、補償したうえで `operationId` と `rolledBack: true` を付ける。
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    DownstreamUnavailable:
      description: |
        下流に接続できない、またはタイムアウトした。`cause` は `DOWNSTREAM_UNAVAILABLE`。`downstream` を付ける。
        書き込みの途中なら、補償したうえで `operationId` と `rolledBack: true` を付ける。
        原因の見当（証明書の不受理、名前解決の失敗など）は、ログと `/status` に出す。
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
  schemas:
...
    Hex128:
      type: string
      pattern: '^[A-Fa-f0-9]{32}$'
...
    Subscriber:
      type: object
      required:
        - imsi
        - keyStore
        - issues
      properties:
        imsi:
          $ref: '#/components/schemas/Imsi'
        keyStore:
          $ref: '#/components/schemas/KeyStore'
        key:
          allOf:
            - $ref: '#/components/schemas/SubscriberKey'
          description: 置き場所の加入者。ない場合は省略する（`issues` に `KEY_MISSING`）。
        policy:
          allOf:
            - $ref: '#/components/schemas/PolicyBody'
          description: 認可ポリシー。ない場合は省略する（`issues` に `POLICY_MISSING`）。
        status:
          allOf:
            - $ref: '#/components/schemas/PolicyStatus'
          description: |
            加入者の状態（認可ポリシーの `status`。0.3.0 で追加）。認可ポリシーがない場合は省略する。
            変更は `PUT /policies/{imsi}/status` で行う。
        issues:
          type: array
          description: 食い違い。正常なら空の配列。
          items:
            $ref: '#/components/schemas/SubscriberIssue'
      example:
        imsi: '001020000000001'
        keyStore: aka
        key:
          amf: '8000'
          sqn: '000000000020'
          sqnType: inc32
          allowPlain: false
          allowedClientIds: [1]
          createdAt: '2026-10-10T00:00:00Z'
          updatedAt: '2026-10-10T00:00:00Z'
        policy:
          default: allow
          rules: []
        status: active
        issues: []
    SubscriberCreate:
      type: object
      required:
        - imsi
        - ki
        - opc
        - policy
      properties:
        imsi:
          $ref: '#/components/schemas/Imsi'
        keyStore:
          allOf:
            - $ref: '#/components/schemas/KeyStore'
          description: 省略すれば PLMN マップから決める。指定した場合、PLMN マップから決まるものと違えば 400（`KEY_STORE_MISMATCH`）。
        ki:
          $ref: '#/components/schemas/Hex128'
        opc:
          $ref: '#/components/schemas/Hex128'
        amf:
          allOf:
            - $ref: '#/components/schemas/Amf'
          default: '8000'
        sqn:
          allOf:
            - $ref: '#/components/schemas/Sqn'
          default: '000000000000'
        sqnType:
          allOf:
            - $ref: '#/components/schemas/SqnType'
          description: '`aka` のときだけ指定できる。省略時は aka-only-server の既定（`inc32`）。'
        allowPlain:
          type: boolean
          description: '`aka` のときだけ指定できる。省略時は `false`。'
        policy:
          allOf:
            - $ref: '#/components/schemas/PolicyBody'
          description: 認可ポリシー。状態（`status`）は指定できず、常に `active` で作る。
...
    SubscriberKeys:
      type: object
      required:
        - ki
        - opc
      properties:
        ki:
          $ref: '#/components/schemas/Hex128'
        opc:
          $ref: '#/components/schemas/Hex128'
    SubscriberList:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Subscriber'
        nextCursor:
          type: string
          description: 次のページがある場合だけ返す（そのページの最後の IMSI）。

    # ---- 操作の記録 ----
...